removed bogus test, moved other test into their own test file
